Django-flights-management/requirements.txt
django
django-environ
mysqlclient
django-cors
django-cors-headers
djangorestframework
djangorestframework-simplejwt
# Used for sending requests to microservice
requests
# I just like clearing cache bro, leave it alone
django_clearcache
pycountry
pytest